feat: add additional validations to protect routes

+++ server/src/resources/questions/question.router.ts
@@ -55,9 +55,14 @@ router.route('/questions/:questionId').get(async (req: e.Request, res: e.Respons
 
 router.route('users/:userId/questions/:questionId').delete(validateSession, async (req: e.Request, res: e.Response, next: e.NextFunction) => {
     try {
-        const { questionId } = req.params;
+        const { questionId, userId } = req.params;
         if (!await questionService.exists({ _id: questionId })) {
             return res.status(404).json({ error: 'Question doesn\'t exist' });
+        }
+        const user = await userService.getOne({ _id: userId });
+        const question = await questionService.getOneById({ _id: questionId });
+        if (user._id.toString() !== question.user.toString()) {
+            return res.status(403).json( { status: 'You can only delete questions posted by you' });
         }
         await questionService.deleteOneById();
         return res.status(200).json({ status: 'Question deleted' });
@@ -82,11 +87,14 @@ router.route('/questions/:questionId/answers').put(validateSession, async (req: 
     }
 });
 
-router.route('/questions/:questionId/answers/:answerId').put(validateSession, async (req: e.Request, res: e.Response, next: e.NextFunction) => {
+router.route('/users/:userId/questions/:questionId/answers/:answerId').put(validateSession, async (req: e.Request, res: e.Response, next: e.NextFunction) => {
     try {
-        const { questionId, answerId } = req.params;
+        const { userId, questionId, answerId } = req.params;
         const scoreUpdateDirection = req.query['score'];
         const isBest = req.query['isBest'];
+        if (!await userService.exists({ _id: userId })) {
+            return res.status(404).json({ error: 'User doesn\'t exist' });
+        }
         if (!await questionService.exists({ _id: questionId })) {
             return res.status(404).json({ error: 'Question doesn\'t exist' });
         }
@@ -94,6 +102,11 @@ router.route('/questions/:questionId/answers/:answerId').put(validateSession, as
             await questionService.updateAnswerScore(questionId, answerId, scoreUpdateDirection);
         }
         if (isBest) {
+            const user = await userService.getOne({ _id: userId });
+            const question = await questionService.getOneById({ _id: questionId });
+            if (user._id.toString() !== question.user.toString()) {
+                return res.status(403).json({ error: 'Question wasn\'t posted by you' })
+            }
             await questionService.setAsBest(questionId, answerId, isBest);
         }
         return res.status(200).json({ status: 'Updated successfully' });
@@ -101,5 +114,6 @@ router.route('/questions/:questionId/answers/:answerId').put(validateSession, as
         next(e);
     }
 });
+
 
 module.exports = router;
